Coda.Algebra

Documentation

class Monoid a => Group a where

a <> inv a = mempty = inv a <> a
inv (inv a) = a

Minimal complete definition

inv

Methods

inv :: a -> a

class (Ord a, Semigroup a) => OrderedSemigroup a

x <= y implies z <> x <= z <> y and x <> z <= y' <>' z

class (OrderedSemigroup a, Monoid a) => OrderedMonoid a

pushout

class (OrderedMonoid a, Group a) => OrderedGroup a

pushout

class OrderedSemigroup a => StrictlyOrderedSemigroup a

x < y implies z <> x < z <> y and x <> z < y' <>' z x '=' y implies z <> x '=' z <> y and x <> z '=' y' <>' z

class (OrderedMonoid a, StrictlyOrderedSemigroup a) => StrictlyOrderedMonoid a

pushout

class (OrderedGroup a, StrictlyOrderedMonoid a) => StrictlyOrderedGroup a

pushout

class Semigroup a => SemigroupAction a b where

act (a <> b) = act a . act b

Minimal complete definition

act

Methods

act :: a -> b -> b

class (Monoid a, SemigroupAction a b) => MonoidAction a b

act mempty = id

class (Group a, MonoidAction a b) => GroupAction a b

class (SemigroupAction a b, Ord b) => MonotoneAction a b

act is strictly monotone in the second argument

The action of any given element of the semigroup is a monotone function

x y == act a x <= act a y

class MonotoneAction a b => StrictlyMonotoneAction a b

act is strictly monotone in its second argument

The action of any given element of the semigroup is a strictly monotone function

x < y implies act a x < act a y x = y implies act a x = act a y

class (SemigroupAction a b, OrderedSemigroup a, Ord b) => OrderedMonotoneAction a b

'act is monotone in first argument

a <= b implies act a x <= act b x

class OrderedMonotoneAction a b => OrderedStrictlyMonotoneAction a b

'act is strictly monotone in first argument

a = b implies act a x = act b x a < b implies act a x < act b x

class (SemigroupAction a b, Semigroup b) => DistributiveAction a b

act a (b <> c) = act a b <> act a c

class (DistributiveAction a b, Monoid b) => UnitaryAction a b

act a mempty = mempty

Lemma "inverse action": If a and b are groups, then inv (act a b) = act a (inv b) Proof:

act a (inv b) <> act a b 
= act a (inv b <> b) -- distributive action
= act a mempty -- left identity
= mempty -- unitary action

act a b <> act a (inv b)
= act a (b <> inv b) -- distributive action
= act a mempty -- right identity
= mempty -- unitary action

Hence, act a (inv b) is an inverse of act a b, and inverses are unique.

Therefore, inv (act a b) = act a (inv b).

class (DistributiveAction a b, DistributiveAction a c, DistributiveAction b c) => DistributiveActionAction a b c

act a (act b c) = act (act a b) (act a c)

Hint: Compare the law for how these distributive actions relate to the law for a single distributive action

data Semi a b

strict semidirect product

Constructors

Semi !a !b