Coda.Message.Language

Description

Language Server Protocol

Cancellation Support

pattern CancelRequest :: Id -> Notification_

Note: You should still reply to the item in question

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#-cancellation-support

_CancelRequest :: Prism' Notification_ Id

pattern RequestCancelled :: ErrorCode

cancelledResponse :: Id -> Response Value Id Value

Language Server Protocol

type DocumentUri = Text

| https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#uri

data Position

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#position

interface Position {
  line: number;
  character: number;
}

Constructors

Position

Fields

positionLine :: !Int

0-based line number

positionCharacter :: !Int

0-based count of utf-16 words (not code-points!)

data Range

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#range

interface Range {
  start: Position;
  end: Position;
}

Constructors

Range

Fields

_rangeStart :: !Position

_rangeEnd :: !Position

data Location

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#location

interface Location {
  uri: DocumentUri;
  range: Range;
}

Constructors

Location

Fields

locationUri :: DocumentUri

locationRange :: Range

data Diagnostic

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#diagnostic

interface Diagnostic {
  range: Range;
  severity?: number;
  code?: number | string;
  source?: string;
  message: string;
}

Constructors

Diagnostic

Fields

diagnosticRange :: !Range

diagnosticSeverity :: !(Maybe Severity)

diagnosticCode :: !(Maybe Id)

diagnosticSource :: !(Maybe Text)

diagnosticMessage :: !Text

data Command a

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#command

interface Command {
  title : string;
  command: string;
  arguments?: any[]
}

Constructors

Command

Fields

commandTitle :: Text

commandCommand :: Text

commandArguments :: Maybe [a]

type Command_ = Command Value

data TextEdit

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textedit

interface TextEdit {
  range: Range;
  newText: string;
}

Constructors

TextEdit

Fields

textEditRange :: Range

textEditNewText :: Text

newtype TextDocumentIdentifier

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textdocumentidentifier

interface TextDocumentIdentifier {
  uri : DocumentUri
}

Constructors

TextDocumentIdentifier

Fields

textDocumentIdentifierUri :: DocumentUri

data VersionedTextDocumentIdentifier

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#versionedtextdocumentidentifier

interface VersionedTextDocumentIdentifier extends TextDocumentIdentifier {
  version: number;
}

Constructors

VersionedTextDocumentIdentifier

Fields

versionedTextDocumentIdentifierUri :: !DocumentUri

versionedTextDocumentIdentifierVersion :: !Int

data TextDocumentItem

An item to transfer a text document from the client to the server.

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textdocumentitem

interface TextDocumentItem {
  uri: DocumentUri;
  languageId: string;
  version: number;
  text: string;
}

Constructors

TextDocumentItem

Fields

textDocumentItemUri :: !DocumentUri

textDocumentItemLanguageId :: !Text

textDocumentItemVersion :: !Int

textDocumentItemText :: !Text

data TextDocumentEdit

Constructors

TextDocumentEdit

Fields

textDocumentEditTextDocument :: !VersionedTextDocumentIdentifier

textDocumentEditEdits :: [TextEdit]

newtype UriMap a

Constructors

UriMap

Fields

runUriMap :: HashMap Text a

data WorkspaceEdit

Constructors

WorkspaceEdit

Fields

workspaceEditChanges :: !(Maybe (UriMap [TextEdit]))

workspaceEditDocumentChanges :: !(Maybe [TextDocumentEdit])

Overloading

class HasRange t where

Minimal complete definition

range

Methods

range :: Lens' t Range

class HasUri t where

Minimal complete definition

uri

Methods

uri :: Lens' t DocumentUri

class HasVersion t where

Minimal complete definition

version

Methods

version :: Lens' t Int

class HasUri t => HasTextDocumentIdentifier t where

Law:

uri = textDocumentIdentifier . uri

Minimal complete definition

textDocumentIdentifier

Methods

textDocumentIdentifier :: Lens' t TextDocumentIdentifier

class (HasVersion t, HasTextDocumentIdentifier t) => HasVersionedTextDocumentIdentifier t where

Laws:

textDocumentIdentifier = versionedTextDocumentIdentifier . textDocumentIdentifier
version = versionedTextDocumentIdentifier . version

Minimal complete definition

versionedTextDocumentIdentifier

Methods

versionedTextDocumentIdentifier :: Lens' t VersionedTextDocumentIdentifier