Coda.Message.Parser

Description

JSON-RPC 2.0 message parsing

Documentation

newtype Parser a

This parser consumes lazy bytestrings

Constructors

Parser

Fields

runParser :: ByteString -> ParseResult a

data ParseResult a

The result of parsing

Constructors

Err

OK !a !ByteString

message :: Parser ByteString

This parses a JSON-RPC 2.0 message

This stops before we get to actually decoding the JSON message.

decodeMessage :: FromJSON a => Parser (Maybe a)

This decodes a JSON-RPC 2.0 message lazily

If the outer parser fails, then the message stream is unrecoverable. If decoding fails, we simply failed to read this message.

decodeMessage' :: FromJSON a => Parser (Maybe a)

This decodes a JSON-RPC 2.0 message eager

If the outer parser fails, then the message stream is unrecoverable. If decoding fails, we simply failed to read this message.

eitherDecodeMessage :: FromJSON a => Parser (Either String a)

This decodes a JSON-RPC 2.0 message lazily with an error message on failure

eitherDecodeMessage' :: FromJSON a => Parser (Either String a)

This decodes a JSON-RPC 2.0 message eager with an error message on failure