Coda.Relative.Class

Description

In order to work with fragments of syntax trees without having to pay an inordinate cost moving around errors and terms, we need a notion of Relative data types that permit relocation cheaply.

Note: There is structure to these classes: Consider the following problematic derivation that show some instances here are incompatible with others.

Delta 1 =                             -- by instance Monoid Delta
Delta 1 <> mempty =               -- by instance MonoidalDelta t
Delta 1 <> delta mempty =       -- by instance RelativeDelta t
delta (rel (Delta 1) mempty) =  -- by instance RelativeMonoid t
delta mempty =                      -- by instance MonoidalDelta t
mempty =                              -- by definition
Delta 0

Data types with relative positions

class Relative a where

Applying a relative position change as a left monoid action

Laws:

rel mempty ≡ id
rel (m <> n) ≡ rel m . rel n

Preferably rel relocates in O(1) or logarithmic time at worst or the container probably isn't well suited for relative use.

Note: rel d = id is a perfectly legitimate definition by these laws.

Note: if you use some stashed delta to slow the descent into your data structure, then you probably need to have nominal roles for your arguments.

Methods

rel :: Delta -> a -> a

rel :: (Generic a, GRelative (Rep a)) => Delta -> a -> a

class GRelative f

Minimal complete definition

grel'

grel :: (Generic a, GRelative (Rep a)) => Delta -> a -> a

We can derive relativity generically.

frel :: (Functor f, Relative a) => Delta -> f a -> f a

Every functor can be relative.

birel :: (Bifunctor f, Relative a, Relative b) => Delta -> f a b -> f a b

Every bifunctor can be relative.

Relative monoids

class (Relative t, Monoid t) => RelativeMonoid t

Laws: rel d is a monoid homomorphism.

rel d (m <> n)   = rel d m <> rel d n
rel d mempty   = mempty

TODO: Add RelativeSemigroup in ghc 8.4

instance RelativeSemigroup Void
instance Relative a => RelativeSemigroup (NonEmpty a)

Relative orderings

class (Relative t, Ord t) => RelativeOrder t

A relative order

Laws:

rel d is monotone, that is to say

x <= y implies rel d x <= rel d y

class RelativeOrder t => StrictRelativeOrder t

A _strict_ relative order

Laws:

x < y implies rel d x < rel d y

This can be useful for keys in relative maps because shifting the map can't cause keys to collapse