Coda.Relative.Delta

Description

Stuff we an measure in UTF-16 code units

Documentation

newtype Delta

A count of UTF-16 code-units.

This forms an (obvious) Abelian group unlike the merely monoidal pairs of line and column.

It is also very compact fitting in a single Int.

Constructors

Delta Int

class HasDelta t where

Something we can measure.

Minimal complete definition

delta

Methods

delta :: t -> Delta

units :: HasDelta t => t -> Int

extract the number of utf-16 code units from a delta

class (Monoid t, HasDelta t) => HasMonoidalDelta t

delta for this type is a monoid homomorphism

delta (m <> n) = delta m <> delta n
delta mempty = 0

class (Ord t, HasDelta t) => HasOrderedDelta t

Requires that delta is monotone

m <= n implies delta m <= delta n