Coda.Relative.Delta.Type

Description

Stuff we an measure in UTF-16 code units

Documentation

newtype Delta

A count of UTF-16 code-units.

This forms an (obvious) Abelian group unlike the merely monoidal pairs of line and column.

It is also very compact fitting in a single Int.

Constructors

Delta Int