Coda.Syntax.Change

Documentation

class Semigroup a => InverseSemigroup a where

a <> inv a <> a = a
inv a <> a <> inv a = inv a

and all idempotents commute and are of the form (a <> inv a) for some a

Minimal complete definition

inv

Methods

inv :: a -> a

data Error

data Partial a

This occupies an uncomfortable middle ground between Maybe and Either

Constructors

OK !a

Fail Error

total :: Partial a -> a

class Composable a where

Methods

compose :: a -> a -> Partial a

compose :: Semigroup a => a -> a -> Partial a

class (Inverse (Idempotent a), Composable a, Idempotent (Idempotent a) ~ Idempotent a) => Inverse a where

Minimal complete definition

inverse

Associated Types

type Idempotent a :: *

Inverse gives us an inverse category, but the types aren't parameters and are down here at the value level

src and tgt are the source and target mappings, and ident provides us the identity arrows.

Methods

inverse :: a -> a

src :: a -> Idempotent a

src :: Idempotent a ~ a => a -> Idempotent a

tgt :: a -> Idempotent a

tgt :: Idempotent a ~ a => a -> Idempotent a

idd :: Idempotent a -> a

idd :: Idempotent a ~ a => Idempotent a -> a

chunky :: Foldable f => f Text -> Text

foldMapWithPos :: forall a m. (Measured a, Monoid m) => (Measure a -> a -> m) -> FingerTree a -> m

class (Measured t, HasDelta (Measure t)) => Splittable t where

Minimal complete definition

splitDelta | takeDelta, dropDelta

Methods

splitDelta :: Delta -> t -> (t, t)

takeDelta :: Delta -> t -> t

dropDelta :: Delta -> t -> t

consText :: Text -> FingerTree Text -> FingerTree Text

snocText :: FingerTree Text -> Text -> FingerTree Text

class Strippable a where

Minimal complete definition

stripSuffixes

Methods

stripSuffixes :: FingerTree Text -> a -> Partial a

class PP a where

Minimal complete definition

pp

Methods

pp :: a -> (String, String)

ppBar :: (String, String)

pad :: Delta -> Char -> String

pretty :: PP a => a -> IO ()

data Grade

Constructors

Grade !Delta !Delta

data Edit

Constructors

Edit !Delta !Delta !Delta

inverseEdit :: Edit -> Edit

class Editable a where

edit e >=> edit (inverseEdit e) >=> edit e = edit e
edit (inverseEdit e) >=> edit e >=> edit (inverseEdit e) = edit (inverseEdit e)

Minimal complete definition

edit

Methods

edit :: Edit -> a -> Partial a

class FromEdit a where

Minimal complete definition

fromEdit

Methods

fromEdit :: Edit -> a

ins :: FromEdit a => Delta -> a

del :: FromEdit a => Delta -> a

cpy :: FromEdit a => Delta -> a

data Change

Invariants:

1) no two edits with 0 spaces between them. they get coalesced into a single edit node

2) all edits have at least one of the finger-trees non-empty

Changes are simplicial morphisms, monotone functions between finite sets of integers that start at 0

Constructors

Change !(FingerTree Edit) !Delta

changePattern :: Change -> (FingerTree Edit, Delta)

pattern C0 :: Delta -> Change

pattern CN :: Edit -> FingerTree Edit -> Delta -> Change

newtype App f a

Constructors

App

Fields

runApp :: f a

class Changeable a where

Minimal complete definition

change

Methods

change :: Change -> a -> Partial a

generalize :: Change -> Change

build a strictly more general function that produces the same answer on all accepted inputs.

generalize is idempotent