Coda.Syntax.Dyck

Description

Documentation

data Dyck

Dyck l ms r s k e

k indicates if the last token was a layout keyword, and if so can provide a numerical indicator as to which one. 0 means either no it wasn't or that we haven't seen a

Constructors

Dyck !(Cat Closing) !(Cat Token) !(Rev Cat Opening) !(Cat Token) !LayoutMode !(Cat MismatchError)

data Opening

Constructors

Opening !LocatedPair !(Cat Token)

data Closing

Constructors

Closing !(Cat Token) !LocatedPair

type LocatedPair = Located Pair

data MismatchError

Constructors

MismatchError !LocatedPair !LocatedPair

token :: Dyck -> Token -> Dyck

O(1)

layoutToken :: Dyck -> LayoutMode -> Token -> Dyck

close :: Dyck -> Located Pair -> Dyck

O(1)

open :: Dyck -> Located Pair -> Dyck

O(1)

spine :: Dyck -> Cat Token