Coda.Syntax.Keywords

Description

Documentation

keywords :: Set String

These are keywords that may occur anywhere in a source file

startingKeywords :: Set String

these are keywords that are only valid at the start of a top level statement

layoutKeywords :: Set String

These are keywords that introduce layout