Coda.Syntax.Line

Lines

newtype Line

Invariants

The only occurrences of '\n', '\r' or "rn" are at the end of the Text

Constructors

Line Text

class HasLine t where

Minimal complete definition

line

Methods

line :: t -> Line

Summarizing Lines

data LineMeasure

LineMeasure is not Position

This measures line count and total number of code-units over all lines not the number of code-units on the current line.

This makes the result a proper Abelian group, not just a Monoid.

Also when used to measure a FingerTree you can use the FingerTree to convert back and forth between Delta and Position in O(log l) time, while still letting us use the compact single-integer Abelian group Delta representation internally for almost all positioning.

Constructors

LineMeasure !Int !Int

class HasDelta t => HasLineMeasure t where

Minimal complete definition

lineCount

Methods

lineCount :: t -> Int

Position - Delta

positionToDelta :: (Measured v a, HasLineMeasure v) => FingerTree v a -> Position -> Delta

Convert from a Language Server Protocol Position to a Delta given the associated text.

O(log l) where l is the number of lines

deltaToPosition :: (Measured v a, HasLineMeasure v) => FingerTree v a -> Delta -> Position

Compute an Language Server Protocol Position from a Delta given the associated text

Takes O(log l) where l is the number of lines