Coda.Syntax.Rope

Documentation

newtype Rope v a

Constructors

Rope (FingerTree (LineMeasure v) (Line a))

fromText :: FromText a => Text -> a

splitAtPosition :: (RelativeMonoid v, Measured v a, FromText a) => Position -> Rope v a -> (Rope v a, Rope v a)

splitAtDelta :: (RelativeMonoid v, Measured v a, FromText a) => Delta -> Rope v a -> (Rope v a, Rope v a)

insertAt :: (RelativeMonoid v, Measured v a, FromText a) => Position -> Text -> Rope v a -> Rope v a

deleteRange :: (RelativeMonoid v, Measured v a, FromText a) => Range -> Rope v a -> Rope v a

replaceRange :: (RelativeMonoid v, Measured v a, FromText a) => Range -> Text -> Rope v a -> Rope v a

deltaToPosition :: (RelativeMonoid v, Measured v a) => Rope v a -> Delta -> Position

Compute an Language Server Protocol Position from a Delta given the associated text

Takes O(log l) where l is the number of lines

positionToDelta :: (RelativeMonoid v, Measured v a) => Rope v a -> Position -> Delta

Convert from a Language Server Protocol Position to a Delta given the associated text.

O(log l) where l is the number of lines