Coda.Syntax.Token

Description

Documentation

data Token

Constructors

Token !Delta !Text

TokenName !Delta !Name

TokenKeyword !Delta !Keyword

TokenInteger !Delta !Integer

TokenDouble !Delta !Double

TokenString !Delta !Text

TokenChar !Delta !Char

TokenNested !(Located Pair) !(Cat Token)

TokenMismatch !(Located Pair) !(Located Pair) !(Cat Token)

TokenUnmatchedOpening !(Located Pair)

TokenUnmatchedClosing !(Located Pair)

TokenLexicalError !Delta String

data Pair

Constructors

Brace

Bracket

Paren

data LayoutMode

Constructors

LNone

LDo

LLet

LOf

LWhere

data Keyword

Constructors

KAs

KCase

KClass

KData

KDefault

KDeriving

KDo

KElse

KHiding

KIf

KImport

KIn

KInfix

KInfixl

KInfixr

KInstance

KLet

KModule

KNewtype

KOf

KQualified

KThen

KType

KWhere

keywords :: Set String

These are keywords that may occur anywhere in a source file

startingKeywords :: Set String

these are keywords that are only valid at the start of a top level statement

layoutKeywords :: Set String

These are keywords that introduce layout

nested :: Located Pair -> Cat Token -> Token

mismatch :: Located Pair -> Located Pair -> Cat Token -> Token

unmatchedOpening :: Located Pair -> Token

unmatchedClosing :: Located Pair -> Token

lexicalError :: Delta -> String -> Token