Coda.Util.Aeson

Description

This module provides a couple of ad hoc combinators that make it easier to share code (at least in style) between toEncoding and toJSON definitions for ToJSON.

Documentation

pattern JSON :: forall a t. (FromJSON a, ToJSON a, AsJSON t) => a -> t

pattern Value_ :: forall a. (FromJSON a, ToJSON a) => a -> Value

pattern Number_ :: forall t. AsNumber t => Scientific -> t

pattern Double :: forall t. AsNumber t => Double -> t

pattern Integer :: forall t. AsNumber t => Integer -> t

pattern Integral :: forall t a. (AsNumber t, Integral a) => a -> t

pattern Primitive :: forall t. AsPrimitive t => Primitive -> t

pattern Bool_ :: forall t. AsPrimitive t => Bool -> t

pattern String_ :: forall t. AsPrimitive t => Text -> t

pattern Null_ :: forall t. AsPrimitive t => t