Coda.Util.Async

Documentation

waitAll :: [Async a] -> IO [a]

Wait for a bunch of Async computations

waitAllSTM :: [Async a] -> STM [a]

Wait for a bunch of Async computations in the STM monad

waitAll_ :: [Async a] -> IO ()

Wait for a bunch of Async computations, discarding the results.

waitAllSTM_ :: [Async a] -> STM ()

Wait for a bunch of Async computations in the STM monad, discarding the results.