Coda.Util.BitQueue

Description

Documentation

data BitQueue

data BitQueueB

emptyQB :: BitQueueB

Create an empty bit queue builder. This is represented as a single guard bit in the most significant position.

snocQB :: BitQueueB -> Bool -> BitQueueB

Enqueue a bit. This works by shifting the queue right one bit, then setting the most significant bit as requested.

buildQ :: BitQueueB -> BitQueue

Convert a bit queue builder to a bit queue. This shifts in a new guard bit on the left, and shifts right until the old guard bit falls off.

unconsQ :: BitQueue -> Maybe (Bool, BitQueue)

Dequeue an element, or discover the queue is empty.

toListQ :: BitQueue -> [Bool]

Convert a bit queue to a list of bits by unconsing. This is used to test that the queue functions properly.