Language.Server.Protocol

Description

JSON-RPC 2.0 and the Language Server Protocol

JSON-RPC 2.0

data Id

A JSON-RPC message identifier

Constructors

IntId !Int

TextId !Text

Requests

data Request

http://www.jsonrpc.org/specification#request_object

interface RequestMessage extends Message {
  id: number | string; -- missing id is a notification, both are unified here
  method: string;
  params?: any
}

Constructors

Request !(Maybe Id) !Text !(Maybe Value)

Reponses

data Response

http://www.jsonrpc.org/specification#response_object

interface ResponseMessage extends Message {
  id: number | string | null;
  result?: any;
  error?: ResponseErrorany;
}

Constructors

Response !(Maybe Id) !(Maybe Value) !(Maybe ResponseError)

data ResponseError

http://www.jsonrpc.org/specification#error_object

interface ResponseErrorD {
  code: number;
  message: string;
  data?: D;
}

Constructors

ResponseError !ErrorCode !Text !(Maybe Value)

Error Codes

newtype ErrorCode

http://www.jsonrpc.org/specification#error_object

Constructors

ErrorCode Int

pattern ParseError :: ErrorCode

pattern InvalidRequest :: ErrorCode

pattern MethodNotFound :: ErrorCode

pattern InvalidParams :: ErrorCode

pattern InternalError :: ErrorCode

pattern ServerErrorStart :: ErrorCode

pattern ServerErrorEnd :: ErrorCode

pattern ServerNotInitialized :: ErrorCode

pattern UnknownErrorCode :: ErrorCode

Language Server Protocol

Cancellation Support

pattern CancelRequest :: Id -> Request

Note: You should still reply to the item in question

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#-cancellation-support

pattern RequestCancelled :: ErrorCode

cancelledResponse :: Id -> Response

Data Types

DocumentUris

newtype DocumentUri

| https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#uri

Constructors

DocumentUri Text

pattern File :: FilePath -> DocumentUri

Encodedecode FilePaths - "file:/" urls

Positions

data Position

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#position

interface Position {
  line: number;
  character: number;
}

Constructors

Position !Int !Int

subtractPosition :: Position -> Position -> Position

subtractPosition a b subtracts a from b.

Be careful, as this is the opposite argument order from (-), but matches the argument order of subtract

This is a clamped subtraction, much as if we defined subtraction on natural numbers to be total by defining n - m | m >= n = 0

Because of the necessary clamping, Position does not form a group.

data Range

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#range

interface Range {
  start: Position;
  end: Position;
}

Constructors

Range !Position !Position

rangeSize :: Range -> Position

Compute the relative position of the end of the range from the start of the range

data Location

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#location

interface Location {
  uri: DocumentUri;
  range: Range;
}

Constructors

Location DocumentUri Range

Diagnostics

data Diagnostic

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#diagnostic

interface Diagnostic {
  range: Range;
  severity?: number;
  code?: number | string;
  source?: string;
  message: string;
}

Constructors

Diagnostic !Range !(Maybe Severity) !(Maybe Id) !(Maybe Text) !Text

newtype Severity

See DiagnosticSeverity in

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#diagnostic

Constructors

Severity Int

pattern Information :: Severity

Reports information

pattern Hint :: Severity

Reports a hint

pattern Warning :: Severity

Reports a warning.

pattern Error :: Severity

Reports an error.

data Command

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#command

interface Command {
  title : string;
  command: string;
  arguments?: any[]
}

Constructors

Command Text Text (Maybe [Value])

data TextEdit

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textedit

interface TextEdit {
  range: Range;
  newText: string;
}

Constructors

TextEdit !Range !Text

newtype TextDocumentIdentifier

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textdocumentidentifier

interface TextDocumentIdentifier {
  uri : DocumentUri
}

Constructors

TextDocumentIdentifier DocumentUri

data VersionedTextDocumentIdentifier

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#versionedtextdocumentidentifier

interface VersionedTextDocumentIdentifier extends TextDocumentIdentifier {
  version: number;
}

Constructors

VersionedTextDocumentIdentifier !DocumentUri !Int

Contents

data TextDocumentItem

An item to transfer a text document from the client to the server.

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textdocumentitem

interface TextDocumentItem {
  uri: DocumentUri;
  languageId: string;
  version: number;
  text: string;
}

Constructors

TextDocumentItem !DocumentUri !Text !Int !Text

data TextDocumentEdit

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#new-textdocumentedit

export interface TextDocumentEdit {
  textDocument: VersionedTextDocumentIdentifier;
  edits: TextEdit[];
}

Constructors

TextDocumentEdit !VersionedTextDocumentIdentifier [TextEdit]

data WorkspaceEdit

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#workspaceedit

export interface WorkspaceEdit {
  changes?: { [uri: string]: TextEdit[]; };
  documentChanges?: TextDocumentEdit[];
}

Constructors

WorkspaceEdit !(Maybe (HashMap Text [TextEdit])) !(Maybe [TextDocumentEdit])

data DocumentFilter

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#documentfilter

 export interface DocumentFilter {
	language?: string;
	scheme?: string;
	pattern?: string;
 }
 

Constructors

DocumentFilter (Maybe String) (Maybe String) (Maybe String)

data TextDocumentPositionParams

https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textdocumentpositionparams

Constructors

TextDocumentPositionParams TextDocumentIdentifier Position

type TDPP = TextDocumentPositionParams

Protocol

initialize

pattern Initialize :: Id -> InitializeParams -> Request

initialize

data InitializeParams

Constructors

InitializeParams (Maybe Int) (Maybe (Maybe String)) (Maybe DocumentUri) (Maybe Value) ClientCapabilities Trace

data ClientCapabilities

Constructors

ClientCapabilities (Maybe WorkspaceClientCapabilities) (Maybe TextDocumentClientCapabilities) (Maybe Value)

type WorkspaceClientCapabilities = Value

type TextDocumentClientCapabilities = Value

initialized

pattern Initialized :: Request

initialized

exit

pattern Exit :: Request

exit

shutdown

pattern Shutdown :: Request

shutdown

'window/logMessage'

pattern LogMessage :: Severity -> Text -> Request

window/logMessage

data LogMessageParams

Constructors

LogMessageParams !Severity !Text

'window/showMessage'

pattern ShowMessage :: Severity -> Text -> Request

window/showMessage

data ShowMessageParams

Constructors

ShowMessageParams !Severity !Text

'telemetry/event'

pattern TelemetryEvent :: Value -> Request

telemetry/event

'client/registerCapability'

pattern RegisterCapability :: Id -> [Registration] -> Request

client/registerCapability

data Registration

Constructors

Registration !Text !Text !(Maybe Value)

newtype TextDocumentRegistrationOptions

Constructors

TextDocumentRegistrationOptions (Maybe [DocumentFilter])

pattern RegisterCapabilityResponse :: Id -> Maybe ResponseError -> Response

'client/unregisterCapability'

pattern UnregisterCapability :: Id -> [Unregistration] -> Request

client/unregisterCapability

data Unregistration

Constructors

Unregistration !Text !Text

'workspace/didChangeConfiguration'

pattern DidChangeConfiguration :: Value -> Request

workspace/didChangeConfiguration

newtype DidChangeConfigurationParams

Constructors

DidChangeConfigurationParams Value

'textDocument/didOpen'

pattern DidOpen :: TextDocumentItem -> Request

textDocument/didOpen

newtype DidOpenTextDocumentParams

Constructors

DidOpenTextDocumentParams TextDocumentItem

'textDocument/didChange'

pattern DidChange :: DidChangeTextDocumentParams -> Request

textDocument/didChange

data DidChangeTextDocumentParams

Constructors

DidChangeTextDocumentParams !VersionedTextDocumentIdentifier [TextDocumentContentChangeEvent]

data TextDocumentContentChangeEvent

Constructors

TextDocumentContentChangeEvent !(Maybe Range) !(Maybe Int) !Text

'textDocument/didClose'

pattern DidClose :: TextDocumentItem -> Request

textDocument/didClose

newtype DidCloseTextDocumentParams

Constructors

DidCloseTextDocumentParams TextDocumentItem

'workspace/didChangeWatchedFiles

pattern DidChangeWatchedFiles :: [FileEvent] -> Request

workspace/didChangeWatchedFiles

newtype DidChangeWatchedFilesParams

Constructors

DidChangeWatchedFilesParams [FileEvent]

data FileEvent

Constructors

FileEvent !DocumentUri !FileChangeType

'textDocument/publishDiagnostics'

pattern PublishDiagnostics :: PublishDiagnosticsParams -> Request

textDocument/publishDiagnostics

data PublishDiagnosticsParams

Constructors

PublishDiagnosticsParams !DocumentUri [Diagnostic]

'textDocument/hover'

pattern Hover :: Id -> TextDocumentPositionParams -> Request

data HoverResult

Constructors

HoverResult [MarkedString] (Maybe Range)

data MarkedString

Constructors

MarkedString !(Maybe Text) !Text

Ad-hoc Overloading

class HasArguments s a | s -> a where

Minimal complete definition

arguments

Methods

arguments :: Lens' s a

class HasCapabilities s a | s -> a where

Minimal complete definition

capabilities

Methods

capabilities :: Lens' s a

class HasChanges s a | s -> a where

Minimal complete definition

changes

Methods

changes :: Lens' s a

class HasCharacter s a | s -> a where

Minimal complete definition

character

Methods

character :: Lens' s a

class HasCode s a | s -> a where

Minimal complete definition

code

Methods

code :: Lens' s a

class HasCommand s a | s -> a where

Minimal complete definition

command

Methods

command :: Lens' s a

class HasContentChanges s a | s -> a where

Minimal complete definition

contentChanges

Methods

contentChanges :: Lens' s a

class HasContents s a | s -> a where

Minimal complete definition

contents

Methods

contents :: Lens' s a

class HasData s a | s -> a where

Minimal complete definition

data_

Methods

data_ :: Lens' s a

class HasDiagnostics s a | s -> a where

Minimal complete definition

diagnostics

Methods

diagnostics :: Lens' s a

class HasDocumentChanges s a | s -> a where

Minimal complete definition

documentChanges

Methods

documentChanges :: Lens' s a

class HasDocumentSelector s a | s -> a where

Minimal complete definition

documentSelector

Methods

documentSelector :: Lens' s a

class HasEdits s a | s -> a where

Minimal complete definition

edits

Methods

edits :: Lens' s a

class HasEnd s a | s -> a where

Minimal complete definition

end

Methods

end :: Lens' s a

class HasError s a | s -> a where

Minimal complete definition

error

Methods

error :: Lens' s a

class HasExperiment s a | s -> a where

Minimal complete definition

experiment

Methods

experiment :: Lens' s a

class HasId s a | s -> a where

Minimal complete definition

id_

Methods

id_ :: Lens' s a

class HasInitializationOptions s a | s -> a where

Minimal complete definition

initializationOptions

Methods

initializationOptions :: Lens' s a

class HasLanguage s a | s -> a where

Minimal complete definition

language

Methods

language :: Lens' s a

class HasLanguageId s a | s -> a where

Minimal complete definition

languageId

Methods

languageId :: Lens' s a

class HasLine s a | s -> a where

Minimal complete definition

line

Methods

line :: Lens' s a

class HasMessage s a | s -> a where

Minimal complete definition

message

Methods

message :: Lens' s a

class HasMethod s a | s -> a where

Minimal complete definition

method

Methods

method :: Lens' s a

class HasNewText s a | s -> a where

Minimal complete definition

newText

Methods

newText :: Lens' s a

class HasParams s a | s -> a where

Minimal complete definition

params

Methods

params :: Lens' s a

class HasPattern s a | s -> a where

Minimal complete definition

pattern

Methods

pattern :: Lens' s a

class HasPosition s a | s -> a where

Minimal complete definition

position

Methods

position :: Lens' s a

class HasProcessId s a | s -> a where

Minimal complete definition

processId

Methods

processId :: Lens' s a

class HasRange s a | s -> a where

Minimal complete definition

range

Methods

range :: Lens' s a

class HasRangeLength s a | s -> a where

Minimal complete definition

rangeLength

Methods

rangeLength :: Lens' s a

class HasRegisterOptions s a | s -> a where

Minimal complete definition

registerOptions

Methods

registerOptions :: Lens' s a

class HasResult s a | s -> a where

Minimal complete definition

result

Methods

result :: Lens' s a

class HasRootPath s a | s -> a where

Minimal complete definition

rootPath

Methods

rootPath :: Lens' s a

class HasRootUri s a | s -> a where

Minimal complete definition

rootUri

Methods

rootUri :: Lens' s a

class HasScheme s a | s -> a where

Minimal complete definition

scheme

Methods

scheme :: Lens' s a

class HasSettings s a | s -> a where

Minimal complete definition

settings

Methods

settings :: Lens' s a

class HasSeverity s a | s -> a where

Minimal complete definition

severity

Methods

severity :: Lens' s a

class HasSource s a | s -> a where

Minimal complete definition

source

Methods

source :: Lens' s a

class HasStart s a | s -> a where

Minimal complete definition

start

Methods

start :: Lens' s a

class HasText s a | s -> a where

Minimal complete definition

text

Methods

text :: Lens' s a

class HasTextDocument s a | s -> a where

Minimal complete definition

textDocument

Methods

textDocument :: Lens' s a

class HasTitle s a | s -> a where

Minimal complete definition

title

Methods

title :: Lens' s a

class HasTrace s a | s -> a where

Minimal complete definition

trace

Methods

trace :: Lens' s a

class HasType s a | s -> a where

Minimal complete definition

type_

Methods

type_ :: Lens' s a

class HasUri s a | s -> a where

Minimal complete definition

uri

Methods

uri :: Lens' s a

class HasValue s a | s -> a where

Minimal complete definition

value

Methods

value :: Lens' s a

class HasVersion s a | s -> a where

Minimal complete definition

version

Methods

version :: Lens' s a

class HasWorkspace s a | s -> a where

Minimal complete definition

workspace

Methods

workspace :: Lens' s a