Coda.Syntax.Dyck

Description

Documentation

data Dyck a

Dyck l ms r s k e

k indicates if the last token was a layout keyword, and if so can provide a numerical indicator as to which one. 0 means either no it wasn't or that we haven't seen a

Constructors

Dyck !(Cat (Closing a)) !(Cat a) !(Rev Cat (Opening a)) !(Cat a) !Mode !(Cat (MismatchError a))

data Opening a

Constructors

Opening !(LocatedPair a) !(Cat a)

data Closing a

Constructors

Closing !(Cat a) !(LocatedPair a)

token :: Relative a => Dyck a -> a -> Dyck a

O(1)

layoutToken :: Relative a => Dyck a -> Mode -> a -> Dyck a

close :: (Eq (Pair a), AsRich a a, Relative a) => Dyck a -> LocatedPair a -> Dyck a

O(1)

open :: Relative a => Dyck a -> LocatedPair a -> Dyck a

O(1)

spine :: (AsRich a a, Relative a) => Dyck a -> Cat a

class Lexer a where

Minimal complete definition

lex

Methods

lex :: Text -> Dyck a