Coda.Syntax.Rope

Description

Documentation

newtype Rope v a

Constructors

Rope (FingerTree (LineMeasure v) (Line a))

splitAtPosition :: (RelativeMonoid v, Measured v a, FromText a) => Position -> Rope v a -> (Rope v a, Rope v a)

splitAtDelta :: (RelativeMonoid v, Measured v a, FromText a) => Delta -> Rope v a -> (Rope v a, Rope v a)

insertAt :: (RelativeMonoid v, Measured v a, FromText a) => Position -> Text -> Rope v a -> Rope v a

deleteRange :: (RelativeMonoid v, Measured v a, FromText a) => Range -> Rope v a -> Rope v a

replaceRange :: (RelativeMonoid v, Measured v a, FromText a) => Range -> Text -> Rope v a -> Rope v a

deltaToPosition :: (RelativeMonoid v, Measured v a) => Rope v a -> Delta -> Position

Compute an Language Server Protocol Position from a Delta given the associated text

Takes O(log l) where l is the number of lines

positionToDelta :: (RelativeMonoid v, Measured v a) => Rope v a -> Position -> Delta

Convert from a Language Server Protocol Position to a Delta given the associated text.

O(log l) where l is the number of lines

Lines

data Line a

Invariants

The only occurrences of '\n', '\r' or "rn" are at the end of the Text

Constructors

Line

Fields

runLine :: !Text

content :: a

foldLines :: (a -> Text -> a) -> a -> Text -> a

class FromText a where

Minimal complete definition

fromText

Methods

fromText :: Text -> a

Summarizing Lines

data LineMeasure v

LineMeasure is not Position

This measures line count and total number of code-units over all lines not the number of code-units on the current line.

This makes the result a proper Abelian group, not just a Monoid.

Also when used to measure a FingerTree you can use the FingerTree to convert back and forth between Delta and Position in O(log l) time, while still letting us use the compact single-integer Abelian group Delta representation internally for almost all positioning.

Constructors

LineMeasure !Int !Delta v

class HasDelta t => HasLineMeasure t v | t -> v where

Minimal complete definition

lineMeasure

Methods

lineMeasure :: t -> LineMeasure v

lineCount :: t -> Int