Coda.Util.Aeson

Description

This module provides a couple of ad hoc combinators that make it easier to share code (at least in style) between toEncoding and toJSON definitions for ToJSON.

Documentation

class Monoid (Ob v) => FromHook v where

Minimal complete definition

(!=)

Associated Types

type Ob v :: *

Methods

(!=) :: Text -> v -> Ob v infixr 8

class Monoid t => ToHook t where

Minimal complete definition

(!~)

Methods

(!~) :: ToJSON v => Text -> v -> t infixr 8

(?~) :: (ToJSON v, ToHook t) => Text -> Maybe v -> t infixr 8

(?=) :: FromHook v => Text -> Maybe v -> Ob v infixr 8

(??~) :: (ToJSON a, ToHook r) => Text -> a -> r infixr 8

parseMissingAsNull :: FromJSON a => Object -> Text -> Parser a