Coda.Util.Aeson

Description

This module provides a couple of ad hoc combinators that make it easier to share code (at least in style) between toEncoding and toJSON definitions for ToJSON.

Documentation

class Monoid (Ob v) => FromHook v where

Minimal complete definition

(!=)

Associated Types

type Ob v :: *

Methods

(!=) :: Text -> v -> Ob v infixr 8

class Monoid t => ToHook t where

Minimal complete definition

(!~)

Methods

(!~) :: ToJSON v => Text -> v -> t infixr 8

(?~) :: (ToJSON v, ToHook t) => Text -> Maybe v -> t infixr 8

(?=) :: FromHook v => Text -> Maybe v -> Ob v infixr 8

(??~) :: (ToJSON a, ToHook r) => Text -> a -> r infixr 8

parseMissingAsNull :: FromJSON a => Object -> Text -> Parser a

pattern synonyms for the lens-aeson prisms

pattern JSON :: forall a t. (FromJSON a, ToJSON a, AsJSON t) => a -> t

pattern Value_ :: forall a. (FromJSON a, ToJSON a) => a -> Value

pattern Number_ :: forall t. AsNumber t => Scientific -> t

pattern Double :: forall t. AsNumber t => Double -> t

pattern Integer :: forall t. AsNumber t => Integer -> t

pattern Integral :: forall t a. (AsNumber t, Integral a) => a -> t

pattern Primitive :: forall t. AsPrimitive t => Primitive -> t

pattern Bool_ :: forall t. AsPrimitive t => Bool -> t

pattern String_ :: forall t. AsPrimitive t => Text -> t

pattern Null_ :: forall t. AsPrimitive t => t