Coda.Util.Aeson

Description

This module provides a couple of ad hoc combinators that make it easier to share code (at least in style) between toEncoding and toJSON definitions for ToJSON.

Documentation

pattern JSON :: forall a t. (FromJSON a, ToJSON a, AsJSON t) => a -> t